# week2/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Advanced Web",
        "title": "API for Advanced Web",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/actors": {
            "get": {
                "description": "Get all actors",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actor"
                ],
                "summary": "Get all actors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Actor"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create an actor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actor"
                ],
                "summary": "Create an actor",
                "parameters": [
                    {
                        "description": "Actor payload",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ActorRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Actor"
                        }
                    }
                }
            }
        },
        "/actors/{id}": {
            "get": {
                "description": "Get an actor",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actor"
                ],
                "summary": "Get an actor",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "actorId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Actor"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an actor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actor"
                ],
                "summary": "Update an actor",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "actorId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Actor payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ActorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Actor"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an actor with the given ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Actor"
                ],
                "summary": "Delete an actor",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "actorId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actor deleted successfully"
                    }
                }
            }
        },
        "/films": {
            "get": {
                "description": "Get all films",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Get all films",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Film"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update a film",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Update a film",
                "parameters": [
                    {
                        "description": "Film payload",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.FilmRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Film"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a film",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Create a film",
                "parameters": [
                    {
                        "description": "Film payload",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.FilmRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Film"
                        }
                    }
                }
            }
        },
        "/films/{id}": {
            "get": {
                "description": "Get a film with the given ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Get a film",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "filmId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Film"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a film with the given ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Film"
                ],
                "summary": "Delete a film",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "filmId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Film deleted successfully"
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Actor": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "lastUpdate": {
                    "type": "string"
                }
            }
        },
        "entity.Film": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "languageId": {
                    "type": "integer"
                },
                "lastUpdate": {
                    "type": "string"
                },
                "length": {
                    "type": "integer"
                },
                "originalLanguageId": {
                    "type": "integer"
                },
                "rating": {
                    "type": "string"
                },
                "releaseYear": {
                    "type": "integer"
                },
                "rentalDuration": {
                    "type": "integer"
                },
                "rentalRate": {
                    "type": "number"
                },
                "replacementCost": {
                    "type": "number"
                },
                "specialFeatures": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.ActorRequest": {
            "type": "object",
            "required": [
                "firstName",
                "lastName"
            ],
            "properties": {
                "firstName": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "lastName": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
        "model.FilmRequest": {
            "type": "object",
            "required": [
                "description",
                "languageId",
                "length",
                "originalLanguageId",
                "rating",
                "releaseYear",
                "rentalDuration",
                "rentalRate",
                "replacementCost",
                "specialFeatures",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 10
                },
                "languageId": {
                    "type": "integer",
                    "minimum": 0
                },
                "length": {
                    "type": "integer",
                    "minimum": 1
                },
                "originalLanguageId": {
                    "type": "integer"
                },
                "rating": {
                    "type": "string"
                },
                "releaseYear": {
                    "type": "integer",
                    "minimum": 0
                },
                "rentalDuration": {
                    "type": "integer",
                    "minimum": 1
                },
                "rentalRate": {
                    "type": "number",
                    "minimum": 0
                },
                "replacementCost": {
                    "type": "number",
                    "minimum": 0
                },
                "specialFeatures": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255
                }
            }
        }
    }
}